+++ apps/netconf/netconf_rpc.c
@@ -453,7 +453,7 @@ netconf_notification_cb(int   s,
 
     clicon_debug(1, "%s", __FUNCTION__);
     /* get msg (this is the reason this function is called) */
-    if (clicon_msg_rcv(s, &reply, &eof) < 0)
+    if (clicon_msg_rcv(s, 0, &reply, &eof) < 0)
         goto done;
     /* handle close from remote end: this will exit the client */
     if (eof){